#935 Migrate Config test contracts to fixtures
