+++ hartshorn-config/src/testFixtures/java/test/org/dockbox/hartshorn/config/SerializationTests.java
