+++ hartshorn-config/src/testFixtures/java/test/org/dockbox/hartshorn/config/mapping/PersistenceModifiersTests.java
@@ -16,7 +16,6 @@
 
 package test.org.dockbox.hartshorn.config.mapping;
 
-import org.dockbox.hartshorn.application.context.ApplicationContext;
 import org.dockbox.hartshorn.config.JsonInclusionRule;
 import org.dockbox.hartshorn.config.ObjectMapper;
 import org.dockbox.hartshorn.config.annotations.UseSerialization;
@@ -28,18 +27,15 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import jakarta.inject.Inject;
-
 @HartshornTest(includeBasePackages = false)
 @UseSerialization
-public class PersistenceModifiersTests {
+public abstract class PersistenceModifiersTests {
 
-    @Inject
-    private ApplicationContext applicationContext;
+    protected abstract ObjectMapper objectMapper();
 
     @Test
     void testSkipEmptyKeepsNonEmpty() {
-        final ObjectMapper mapper = this.applicationContext.get(ObjectMapper.class).skipBehavior(JsonInclusionRule.SKIP_EMPTY);
+        final ObjectMapper mapper = this.objectMapper().skipBehavior(JsonInclusionRule.SKIP_EMPTY);
         final ModifierElement element = new ModifierElement(List.of("sample", "other"));
         final Option<String> out = mapper.write(element);
 
@@ -49,7 +45,7 @@ public class PersistenceModifiersTests {
 
     @Test
     void testSkipEmptySkipsEmpty() {
-        final ObjectMapper mapper = this.applicationContext.get(ObjectMapper.class).skipBehavior(JsonInclusionRule.SKIP_EMPTY);
+        final ObjectMapper mapper = this.objectMapper().skipBehavior(JsonInclusionRule.SKIP_EMPTY);
         final ModifierElement element = new ModifierElement(List.of());
         final Option<String> out = mapper.write(element);
