+++ hartshorn-config/src/testFixtures/resources/junit.properties
